Move analyzer/general.py to analyzer/__init__.py
